+++ src/main/java/com/example/WeBKKHustraUhrovec/Controller/PlayerController.java
@@ -2,8 +2,14 @@ package com.example.WeBKKHustraUhrovec.Controller;
 
 
 import com.example.WeBKKHustraUhrovec.Entity.Player;
+import com.example.WeBKKHustraUhrovec.Entity.User;
+import com.example.WeBKKHustraUhrovec.Enum.UserRole;
 import com.example.WeBKKHustraUhrovec.Service.PlayerService;
+import com.example.WeBKKHustraUhrovec.Service.UserService;
+import com.example.WeBKKHustraUhrovec.jwt.JwtTokenUtil;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -17,13 +23,42 @@ public class PlayerController {
     @Autowired
     private PlayerService playerService;
 
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
+
+    @Autowired
+    private UserService userService;
+
+
     @PostMapping(path = "/save")
-    public Player savePlayer(@RequestParam String id, @RequestBody Player player) {
-        return playerService.addPlayer(id, player);
+    public ResponseEntity<Player> savePlayer(@RequestParam String id,
+                                             @RequestBody Player player,
+                                             @RequestHeader(value = "Authorization", required = false) String token) {
+        if (token == null || jwtTokenUtil.isTokenExpired(token.substring(7))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+        User user = userService.getUserByToken(token.substring(7));
+        if (user.getRole() != UserRole.ADMIN) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
+
+        Player savedPlayer = playerService.addPlayer(id, player);
+        return ResponseEntity.ok(savedPlayer);
     }
+
     @PostMapping(path = "/saveUhrovec")
-    public Player savePlayerUhrovec(@RequestBody Player player) {
-        return playerService.addPlayerUhrovec(player);
+    public ResponseEntity<Player> savePlayerUhrovec(@RequestBody Player player,
+                                                    @RequestHeader(value = "Authorization", required = false) String token) {
+        if (token == null || jwtTokenUtil.isTokenExpired(token.substring(7))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+        User user = userService.getUserByToken(token.substring(7));
+        if (user.getRole() != UserRole.ADMIN) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
+
+        Player savedPlayer = playerService.addPlayerUhrovec(player);
+        return ResponseEntity.ok(savedPlayer);
     }
     @GetMapping(path = "/getPlayer")
     public Optional<Player> getPlayer(@RequestParam String id) {
@@ -48,12 +83,37 @@ public class PlayerController {
     @GetMapping(path = "/getPlayersList")
     public List<Player> getAllPlayers() {return playerService.getAllPlayers();}
     @DeleteMapping(path = "/deletePlayer")
-    public String deletePlayer(@RequestParam Integer id) {
-        return playerService.deletePlayer(id);
+    public ResponseEntity<String> deletePlayer(@RequestParam Integer id,
+                                               @RequestHeader(value = "Authorization", required = false) String token) {
+        if (token == null || jwtTokenUtil.isTokenExpired(token.substring(7))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        }
+        User user = userService.getUserByToken(token.substring(7));
+        if (user.getRole() != UserRole.ADMIN) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
+
+        String result = playerService.deletePlayer(id);
+        return ResponseEntity.ok(result);
     }
+
     @PutMapping(path = "/updatePlayer")
-    public Player updatePlayer(@RequestBody Player player) {
-        return playerService.updatePlayer(player);
+    public ResponseEntity<Player> updatePlayer(@RequestBody Player player,
+                                               @RequestHeader(value = "Authorization", required = false) String token) {
+        if (token == null || jwtTokenUtil.isTokenExpired(token.substring(7))) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+        User user = userService.getUserByToken(token.substring(7));
+        if (user.getRole() != UserRole.ADMIN) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
+
+        Player updatedPlayer = playerService.updatePlayer(player);
+        if (updatedPlayer != null) {
+            return ResponseEntity.ok(updatedPlayer);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
 }
